Moved all columns back into Searchresults. Conditionally render status column
