+++ Client/src/Components/Profile/UserUploadedDatasetsSection/UploadedDatasetsTab.tsx
@@ -1,12 +1,13 @@
-import React, { useEffect, useState } from 'react'
-import { ColDef, DataGrid, SelectionChangeParams, ValueGetterParams } from '@material-ui/data-grid';
 import { Box, Container, Grid, List, ListItem, ListItemText } from '@material-ui/core';
+import { ColDef, DataGrid, SelectionChangeParams, ValueGetterParams } from '@material-ui/data-grid';
 import { IData, IDatasetModel, IDatasetStatus, IReference } from '../../../Models/Datasets/IDatasetModel';
-import { ICategoryModel } from '../../../Models/Profile/ICategoryModel';
+import React, { useEffect, useState } from 'react'
 import { callGetDatasets, getUploadedDatasets } from '../../../Remote/Endpoints/DatasetEndpoint';
-import { listCategories } from '../../../Remote/Endpoints/CategoryEndpoint';
-import { SearchResults } from '../../Search/SearchResults';
+
 import { DatasetFormModal } from '../../DatasetUpload/DatasetViewModal';
+import { ICategoryModel } from '../../../Models/Profile/ICategoryModel';
+import { SearchResults } from '../../Search/SearchResults';
+import { listCategories } from '../../../Remote/Endpoints/CategoryEndpoint';
 
 interface IProps {
     datasetResults?: IDatasetModel[],
@@ -40,14 +41,6 @@ export const UploadedDatasetsTab = () => {
             console.log("datasetStatus: ", status_array)
         }
 
-        const getDatasetStatus = async () => {
-            const uploaded_datasets = await getUploadedDatasets()
-            const status_array = uploaded_datasets.map(x => x.approved)
-            setDatasetStatus(status_array)
-            console.log("status: ", status_array)
-            return status_array
-        }
-
         const callListCategories = async () => {
             const getCategories = await listCategories()
             setCategories(getCategories || [])
@@ -64,83 +57,6 @@ export const UploadedDatasetsTab = () => {
 
     }, [datasetIds])
 
-
-    const getTitle = (params: ValueGetterParams) => {
-        const reference = params.getValue('reference') as IReference
-        return `${reference.title}`
-    }
-
-
-    const getCategoryId = (params: ValueGetterParams) => {
-        const categoryId = params.row.category
-        const foundCategory = categories.find(category => category.id == categoryId)
-        return `${foundCategory.name}`
-    }
-
-
-    const getStatus = (params: ValueGetterParams) => {
-
-        //const status = params.row.status
-        const dataset_id = params.row.status
-        const foundDataset = datasetStatus.find(dataset => dataset.datasetId == dataset_id)
-
-        return `${foundDataset}`
-
-    }
-
-    const getSubcategoryId = (params: ValueGetterParams) => {
-        const subcategoryId = params.row.subcategory
-        const categoryId = params.row.category
-        const foundCategory = categories.find(category => category.id == categoryId)
-        const foundSubcategory = foundCategory.subcategories.find(subcategory => subcategory.id == subcategoryId)
-        return `${foundSubcategory.name}`
-    }
-
-    const getAuthor = (params: ValueGetterParams) => {
-        const reference = params.getValue('reference') as IReference
-        if (!reference.authors[0]) {
-            return 'N/A'
-        }
-        return `${reference.authors[0].firstName} ${reference.authors[0].lastName}`
-    }
-
-    const getYear = (params: ValueGetterParams) => {
-        const reference = params.getValue('reference') as IReference
-        return `${reference.year}`
-    }
-
-    const getVariableList = (params: ValueGetterParams) => {
-        const data = params.getValue('data') as IData
-
-        return (
-            <List dense disablePadding style={{ maxHeight: '100%', overflow: 'auto', width: '100%' }}>
-                {data.variables.map(variable =>
-                    <ListItem key={variable.name} dense disableGutters style={{ paddingTop: '0', paddingBottom: '0' }}>
-                        <ListItemText primary={variable.name} style={{ paddingTop: '0', paddingBottom: '0', marginTop: '0', marginBottom: '0' }} />
-                    </ListItem>
-                )}
-            </List>
-        )
-    }
-    const linkToDataset = (params: ValueGetterParams) => {
-        const datasetId = params.getValue('id')
-
-        return (<DatasetFormModal datasetId={datasetId.toString()} />)
-    }
-
-
-    const columns: ColDef[] = [
-        { field: 'dataset_name', headerName: 'Name', flex: 1 },
-        { field: `title`, headerName: 'Title', valueGetter: getTitle, flex: 1 },
-        { field: 'category', headerName: 'Category', flex: 1, valueGetter: getCategoryId },
-        { field: 'subcategory', headerName: 'SubCategory', flex: 1, valueGetter: getSubcategoryId },
-        { field: 'author', headerName: 'Author', flex: 1, valueGetter: getAuthor },
-        { field: 'year', headerName: 'Year', flex: 1, valueGetter: getYear },
-        { field: 'variables', headerName: 'List Of Variables', flex: 2, renderCell: getVariableList },
-        { field: 'dataset_button', headerName: 'View', flex: 1, renderCell: linkToDataset },
-        { field: 'status', headerName: 'Status', flex: 1, valueGetter: getStatus },
-    ]
-
     return (
         <Container>
             <Box pt={4}>
@@ -148,7 +64,7 @@ export const UploadedDatasetsTab = () => {
                     datasetResults={datasets}
                     categories={categories}
                     displayCheckbox={false}
-                    columns={columns}
+                    datasetStatus={datasetStatus}
                 />
             </Box>
         </Container>
